pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "chaos-messenger"
version = "0.1.0"
description = "Open source messenger with blockchain-style encryption"
readme = "README.md"
requires-python = ">=3.11"
authors = [
    {name = "vifirsanova", email = "vifirsanova@gmail.com"}
]
license = {text = "Apache-2.0"}
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3.11",
    "Framework :: AsyncIO",
    "Topic :: Communications :: Chat",
    "Topic :: Security :: Cryptography"
]

dependencies = [
    "sqlalchemy>=2.0.0",
    "asyncpg>=0.29.0",
    "psycopg2-binary>=2.9.9",
    "python-dotenv>=1.0.0",
    "cryptography>=42.0.0",
    "pynacl>=1.5.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.30.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-env>=1.0.0",
    "alembic>=1.13.0",
    "black>=24.0.0",
    "ruff>=0.3.0",
    "mypy>=1.8.0"
]
test = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-env>=1.0.0",
    "pytest-cov>=5.0.0"
]
docs = [
    "sphinx>=7.2.0",
    "sphinx-rtd-theme>=2.0.0"
]

[project.urls]
Homepage = "https://github.com/vifirsanova/chaos-messenger"
Repository = "https://github.com/vifirsanova/chaos-messenger"
Documentation = "https://github.com/vifirsanova/chaos-messenger"

[tool.setuptools.packages.find]
include = ["app*", "tests*"]
exclude = ["tests.old*", "docs*"]

[tool.pytest.ini_options]
minversion = "8.0"
testpaths = ["tests"]
pythonpath = ["."]
asyncio_mode = "auto"
addopts = "-v --tb=short --strict-markers"
asyncio_default_fixture_loop_scope = "session"

[tool.coverage.run]
source = ["app"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:"
]

[tool.black]
line-length = 100
target-version = ["py311"]
include = '\.pyi?$'
extend-exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "I", "B", "C4", "UP"]
ignore = ["B904", "C408"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
strict_optional = true
disallow_untyped_defs = true
disallow_any_unimported = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true

[[tool.mypy.overrides]]
module = ["cryptography.*", "nacl.*", "asyncpg.*"]
ignore_missing_imports = true
